Support CSS3 property word-wrap when rendering PDF
Very long words without spaces are not wraped, which might result in
words printed over the right edge of the paper when rendering PDFs.

CSS3 introduced the property word-wrap that can be used to force wrapping
of long words even if they don't have any spaces.

This patch adds support for the word-wrap property to PDF rendering.