Fix "page-break-after: avoid" followed by multi-line elements
If a document has two consecutive elements <elem1> and <elem2>, and <elem1>
has the property "page-break-after: avoid", then <elem1> should only be moved
to the next page only if there is a page break between <elem1> and <elem2>.
If <elem2> starts on the same page as <elem1>, then later page breaks inside
of <elem2> should not be considered.

This patch fixes the page break behaviour such that "page-break-after: avoid"
is ignored as long as the first line of <elem2> fits on the same page as
the end of <elem1>.

Example: Our documents have the following structure:

<h1>headline</h1>
<h2>headline</h2>
<table><!-- very long table, might include page breaks --></table>
<h2>headline</h2>
<h3>headline</h3>
<table><!-- very long table, might include page breaks --></table>
<h2>headline</h2>
<table><!-- very long table, might include page breaks --></table>

We want to avoid headlines at the bottom of a page, so we put the following
in the CSS:

h1, h2, h3, h4, h5, h6 {
        page-break-after: avoid;
}

The expected behaviour is: When there is a page break after a headline,
the headline should be moved to the next page. If the table begins on the
same page as the headline, then everything is ok, even if there are page breaks
later in the table.

The current implementation is: headlines are moved to the next page unless
the following table fits entirely on the same page as the headline.

This patch fixes the current implementation, such that only the first lines
of the tables are considered when processing "page-break-after: avoid".