Issue 113: Add initial support for quotes CSS 2.1 property.
It is now possible to specify the quotes property and to
configure the content property to use the quotes with the
identifiers: open-quote, no-open-quote, close-quote, and
no-close-quote.

This initial version does not offer full quotes support.  It
is currently possible only to use the first-level quotes 
enumerated in the quotes list or none to suppress quoting.
The QuotingExample program demonstrates how to configure the
quotes and content properties to elicit the insertion of 
quotes.  The example also demonstrate the current flaw where
embedded quotes do not use quotes for the correct depth.