Add -fs-dynamic-auto-width property
If -fs-dynamic-auto-width: dynamic is set on a block-level element whose
width is solely constrained by the page (that is, it and all its
ancestors have width: auto and are not a shrink-to-fit style), it will
have its width recalculated relative to the current page width.

This allows for more natural sizing of elements across pages of varying
size even though (for example) the width of <body> is always determined
by the width of the first page.